backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
langchain>=0.2.6
langchain-experimental>=0.0.61
langchain-google-genai>=1.0.7
langchain-community>=0.2.6
pandas>=2.2.2
matplotlib>=3.9.0
seaborn>=0.13.2
numpy>=1.26.4
python-dotenv>=1.0.1
aiofiles>=23.2.1
Pillow>=10.3.0
openpyxl>=3.1.4
